Cleaned up versions of lab scripts
