+++ server/labs/model_fit/README.md
@@ -1,0 +1,18 @@
+# MODSIR-19 LABS model_fit Module File
+
+## Introduction
+
+The labs directory (and consequently Python module) contains a series of scripts 
+allowing to conveniently interact with the simulator.py code. 
+
+The model_fit module allows 
+to adjust the parameters of a model to fit over actual measured data points.
+
+## Scripts
+
+Please note scripts should be run using the `-m` flag, like this:
+```
+python3 -m labs.model_fit.script
+```
+
+* `optimise.py`: ... to be done ... please run with `--help` argumen to have a short description.
